// src/app/api/models/vehicle.ts
/* tslint:disable */
export interface Vehicle  {

  /**
   * body type
   */
  bodyType?: string;

  /**
   * engine capacity in cc
   */
  engineCapacity?: number;

  /**
   * engine power in Horsepower units
   */
  enginePowerPS?: number;

  /**
   * engine power in KiloWatts
   */
  enginePowerPW?: number;

  /**
   * fuel type
   */
  fuelType?: string;

  /**
   * the make of the car
   */
  make?: string;

  /**
   * the model of the car
   */
  model?: string;
}